+++ src/main/java/insane96mcp/enhancedai/modules/zombie/entity/projectile/FishingHook.java
@@ -192,7 +192,7 @@ public class FishingHook extends Projectile {
         Entity owner = this.getOwner();
         if (owner != null) {
             //TODO Add some min Y force (1d)
-            Vec3 vec3 = (new Vec3(owner.getX() - this.getX(), Math.max(owner.getY() - this.getY(), 1d), owner.getZ() - this.getZ())).scale(0.35D);
+            Vec3 vec3 = (new Vec3(owner.getX() - this.getX(), Math.max(owner.getY() - this.getY(), 1d), owner.getZ() - this.getZ())).scale(0.3D);
             entity.setDeltaMovement(entity.getDeltaMovement().add(vec3));
         }
     }