+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 1.15.0
 * Added Attacking>Melee Attack Speed Based
   * Makes melee mobs attack rate based off attack speed (reduced by 40% compared to players)
 * Added 'Wearden Interactions'